Delete session/tool_calls.rs; migrate load_tool_call_result
The standalone save_tool_call / update_started_tool_call / complete_tool_call
async functions are replaced by ToolCallLifecycle. The remaining live
function load_tool_call_result is migrated to tool_call_lifecycle/query.rs
where it joins ToolCallLifecycle::load. Tests of the old write-side
functions are removed.
