+++ crates/defra-agent/src/session/tests.rs
@@ -87,360 +87,6 @@ async fn compaction_entries_track_files_cumulatively() {
     let _ = std::fs::remove_dir_all(&data_path);
 }
 
-#[tokio::test]
-async fn complete_tool_call_preserves_started_at_datetime() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "read_file",
-        "call-123",
-        r#"{"path":"src/lib.rs"}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    complete_tool_call(&node, "session-1", "call-123", "file contents", "completed")
-        .await
-        .unwrap();
-
-    let resp = node
-        .execute(
-            r#"{
-                AgentToolCall(
-                    filter: { tool_call_id: { _eq: "call-123" } },
-                    limit: 1
-                ) {
-                    status
-                    result
-                    started_at
-                    completed_at
-                }
-            }"#,
-        )
-        .await;
-    assert!(
-        !resp.has_errors(),
-        "query tool call failed: {:?}",
-        resp.errors
-    );
-
-    let row = resp
-        .data
-        .as_ref()
-        .and_then(|data| data.get("AgentToolCall"))
-        .and_then(|value| value.as_array())
-        .and_then(|rows| rows.first())
-        .cloned()
-        .expect("tool call row");
-
-    assert_eq!(
-        row.get("status").and_then(|value| value.as_str()),
-        Some("completed")
-    );
-    assert_eq!(
-        row.get("result").and_then(|value| value.as_str()),
-        Some("file contents")
-    );
-    assert!(row
-        .get("started_at")
-        .and_then(|value| value.as_str())
-        .is_some_and(|value| !value.is_empty()));
-    assert!(row
-        .get("completed_at")
-        .and_then(|value| value.as_str())
-        .is_some_and(|value| !value.is_empty()));
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
-
-#[tokio::test]
-async fn complete_tool_call_persists_trace_enrichment_for_successful_meta_call() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "call_tool",
-        "call-meta-123",
-        r#"{"service_id":"x-data","tool_name":"search_posts","arguments":{"query":"amy"}}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    complete_tool_call(
-        &node,
-        "session-1",
-        "call-meta-123",
-        "search results",
-        "completed",
-    )
-    .await
-    .unwrap();
-
-    let row = load_tool_call_trace_row(&node, "call-meta-123").await;
-    assert_eq!(
-        row.get("selected_service_id")
-            .and_then(|value| value.as_str()),
-        Some("x-data")
-    );
-    assert_eq!(
-        row.get("selected_tool_name")
-            .and_then(|value| value.as_str()),
-        Some("search_posts")
-    );
-    assert!(row
-        .get("tool_failure_class")
-        .is_some_and(|value| value.is_null()));
-    assert!(row
-        .get("latency_ms")
-        .and_then(|value| value.as_i64())
-        .is_some_and(|value| value >= 0));
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
-
-#[tokio::test]
-async fn complete_tool_call_persists_known_failure_class() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "describe_tool",
-        "call-missing-tool",
-        r#"{"service_id":"x-data","tool_name":"missing"}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    complete_tool_call(
-        &node,
-        "session-1",
-        "call-missing-tool",
-        "tool 'missing' not found on service 'x-data'. Available tools: search_posts",
-        "completed",
-    )
-    .await
-    .unwrap();
-
-    let row = load_tool_call_trace_row(&node, "call-missing-tool").await;
-    assert_eq!(
-        row.get("tool_failure_class")
-            .and_then(|value| value.as_str()),
-        Some("serviceUnavailable")
-    );
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
-
-#[tokio::test]
-async fn save_tool_call_update_preserves_existing_trace_enrichment() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "call_tool",
-        "call-resaved-tool",
-        r#"{"service_id":"x-data","tool_name":"search_posts","arguments":{"query":"amy"}}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "call_tool",
-        "call-resaved-tool",
-        r#"{"service_id":"other-service","tool_name":"other_tool","arguments":{}}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    let row = load_tool_call_trace_row(&node, "call-resaved-tool").await;
-    assert_eq!(
-        row.get("selected_service_id")
-            .and_then(|value| value.as_str()),
-        Some("x-data")
-    );
-    assert_eq!(
-        row.get("selected_tool_name")
-            .and_then(|value| value.as_str()),
-        Some("search_posts")
-    );
-    assert!(row
-        .get("tool_failure_class")
-        .is_some_and(|value| value.is_null()));
-    assert!(row.get("latency_ms").is_some_and(|value| value.is_null()));
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
-
-#[tokio::test]
-async fn save_tool_call_does_not_clobber_completed_trace_enrichment() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "describe_tool",
-        "call-resaved-complete",
-        r#"{"service_id":"x-data","tool_name":"missing"}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    complete_tool_call(
-        &node,
-        "session-1",
-        "call-resaved-complete",
-        "tool 'missing' not found on service 'x-data'. Available tools: search_posts",
-        "completed",
-    )
-    .await
-    .unwrap();
-    let completed_row = load_tool_call_trace_row(&node, "call-resaved-complete").await;
-    let completed_latency = completed_row
-        .get("latency_ms")
-        .and_then(|value| value.as_i64())
-        .expect("completed latency");
-
-    save_tool_call(
-        &node,
-        "session-1",
-        1,
-        "describe_tool",
-        "call-resaved-complete",
-        r#"{"service_id":"other-service","tool_name":"other_tool"}"#,
-        "called",
-    )
-    .await
-    .unwrap();
-
-    let row = load_tool_call_trace_row(&node, "call-resaved-complete").await;
-    assert_eq!(
-        row.get("selected_service_id")
-            .and_then(|value| value.as_str()),
-        Some("x-data")
-    );
-    assert_eq!(
-        row.get("selected_tool_name")
-            .and_then(|value| value.as_str()),
-        Some("missing")
-    );
-    assert_eq!(
-        row.get("tool_failure_class")
-            .and_then(|value| value.as_str()),
-        Some("serviceUnavailable")
-    );
-    assert_eq!(
-        row.get("latency_ms").and_then(|value| value.as_i64()),
-        Some(completed_latency)
-    );
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
-
-#[tokio::test]
-async fn agent_tool_call_trace_enrichment_fields_are_nullable_for_old_rows() {
-    let data_path =
-        std::env::temp_dir().join(format!("agent-daemon-tool-call-{}", uuid::Uuid::new_v4()));
-    let node = defra_node::EmbeddedNode::builder()
-        .data_path(&data_path)
-        .build()
-        .await
-        .unwrap();
-    ensure_schemas(&node).await.unwrap();
-
-    let resp = node
-        .execute(
-            r#"mutation {
-                create_AgentToolCall(input: {
-                    tool_call_key: "session-1:old-call",
-                    session_id: "session-1",
-                    message_sequence: 1,
-                    tool_name: "read_file",
-                    tool_call_id: "old-call",
-                    args: "{\"path\":\"src/lib.rs\"}",
-                    result: "file contents",
-                    status: "completed",
-                    started_at: "2026-05-07T12:00:00Z",
-                    completed_at: "2026-05-07T12:00:01Z"
-                }) { _docID }
-            }"#,
-        )
-        .await;
-    assert!(
-        !resp.has_errors(),
-        "create legacy tool call failed: {:?}",
-        resp.errors
-    );
-
-    let row = load_tool_call_trace_row(&node, "old-call").await;
-    for field in [
-        "selected_service_id",
-        "selected_tool_name",
-        "tool_failure_class",
-        "latency_ms",
-    ] {
-        assert!(
-            row.get(field).is_some_and(|value| value.is_null()),
-            "expected {field} to be null, got {:?}",
-            row.get(field)
-        );
-    }
-
-    let _ = std::fs::remove_dir_all(&data_path);
-}
 
 #[tokio::test]
 async fn close_session_preserves_started_datetime() {
@@ -508,40 +154,6 @@ async fn close_session_preserves_started_datetime() {
     let _ = std::fs::remove_dir_all(&data_path);
 }
 
-async fn load_tool_call_trace_row(
-    node: &defra_node::EmbeddedNode,
-    tool_call_id: &str,
-) -> serde_json::Value {
-    let tool_call_id = escape_graphql_string(tool_call_id);
-    let resp = node
-        .execute(&format!(
-            r#"{{
-                AgentToolCall(
-                    filter: {{ tool_call_id: {{ _eq: "{tool_call_id}" }} }},
-                    limit: 1
-                ) {{
-                    selected_service_id
-                    selected_tool_name
-                    tool_failure_class
-                    latency_ms
-                }}
-            }}"#
-        ))
-        .await;
-    assert!(
-        !resp.has_errors(),
-        "query tool call trace fields failed: {:?}",
-        resp.errors
-    );
-
-    resp.data
-        .as_ref()
-        .and_then(|data| data.get("AgentToolCall"))
-        .and_then(|value| value.as_array())
-        .and_then(|rows| rows.first())
-        .cloned()
-        .expect("tool call trace row")
-}
 
 #[tokio::test]
 async fn create_session_with_id_is_idempotent() {
